+++ selfdrive/hardware/tici/test_power_draw.py
@@ -22,7 +22,7 @@ PROCS = [
   Proc('camerad', 2.17),
   Proc('modeld', 0.95),
   Proc('dmonitoringmodeld', 0.25),
-  Proc('loggerd', 0.45, warmup=10.),
+  Proc('encoderd', 0.42),
 ]
