+++ selfdrive/loggerd/encoder/v4l_encoder.h
@@ -2,7 +2,6 @@
 
 #include "selfdrive/common/queue.h"
 #include "selfdrive/loggerd/encoder/encoder.h"
-#include "selfdrive/loggerd/encoder/video_writer.h"
 
 #define BUF_IN_COUNT 7
 #define BUF_OUT_COUNT 6
