+++ selfdrive/loggerd/loggerd.cc
@@ -1,163 +1,18 @@
 #include "selfdrive/loggerd/loggerd.h"
-#include "selfdrive/loggerd/remote_encoder.h"
-bool env_remote_encoder = getenv("REMOTE_ENCODER") != NULL;
+#include "selfdrive/loggerd/video_writer.h"
 
 ExitHandler do_exit;
 
-// Handle initial encoder syncing by waiting for all encoders to reach the same frame id
-bool sync_encoders(LoggerdState *s, CameraType cam_type, uint32_t frame_id) {
-  if (s->camera_synced[cam_type]) return true;
-
-  if (s->max_waiting > 1 && s->encoders_ready != s->max_waiting) {
-    // add a small margin to the start frame id in case one of the encoders already dropped the next frame
-    update_max_atomic(s->start_frame_id, frame_id + 2);
-    if (std::exchange(s->camera_ready[cam_type], true) == false) {
-      ++s->encoders_ready;
-      LOGD("camera %d encoder ready", cam_type);
-    }
-    return false;
-  } else {
-    if (s->max_waiting == 1) update_max_atomic(s->start_frame_id, frame_id);
-    bool synced = frame_id >= s->start_frame_id;
-    s->camera_synced[cam_type] = synced;
-    if (!synced) LOGD("camera %d waiting for frame %d, cur %d", cam_type, (int)s->start_frame_id, frame_id);
-    return synced;
-  }
-}
-
-bool trigger_rotate_if_needed(LoggerdState *s, int cur_seg, uint32_t frame_id) {
-  const int frames_per_seg = SEGMENT_LENGTH * MAIN_FPS;
-  if (cur_seg >= 0 && frame_id >= ((cur_seg + 1) * frames_per_seg) + s->start_frame_id) {
-    // trigger rotate and wait until the main logger has rotated to the new segment
-    ++s->ready_to_rotate;
-    std::unique_lock lk(s->rotate_lock);
-    s->rotate_cv.wait(lk, [&] {
-      return s->rotate_segment > cur_seg || do_exit;
-    });
-    return !do_exit;
-  }
-  return false;
-}
-
-void encoder_thread(LoggerdState *s, const LogCameraInfo &cam_info) {
-  util::set_thread_name(cam_info.filename);
-
-  int cur_seg = -1;
-  int encode_idx = 0;
-  LoggerHandle *lh = NULL;
-  std::vector<Encoder *> encoders;
-  VisionIpcClient vipc_client = VisionIpcClient("camerad", cam_info.stream_type, false);
-
-  // While we write them right to the log for sync, we also publish the encode idx to the socket
-  const char *service_name = cam_info.type == DriverCam ? "driverEncodeIdx" : (cam_info.type == WideRoadCam ? "wideRoadEncodeIdx" : "roadEncodeIdx");
-  PubMaster pm({service_name});
-
-  while (!do_exit) {
-    if (!vipc_client.connect(false)) {
-      util::sleep_for(5);
-      continue;
-    }
-
-    // init encoders
-    if (encoders.empty()) {
-      VisionBuf buf_info = vipc_client.buffers[0];
-      LOGD("encoder init %dx%d", buf_info.width, buf_info.height);
-
-      // main encoder
-      encoders.push_back(new Encoder(cam_info.filename, cam_info.type, buf_info.width, buf_info.height,
-                                     cam_info.fps, cam_info.bitrate,
-                                     cam_info.is_h265 ? cereal::EncodeIndex::Type::FULL_H_E_V_C : cereal::EncodeIndex::Type::QCAMERA_H264,
-                                     buf_info.width, buf_info.height, cam_info.record));
-      // qcamera encoder
-      if (cam_info.has_qcamera) {
-        encoders.push_back(new Encoder(qcam_info.filename, cam_info.type, buf_info.width, buf_info.height,
-                                       qcam_info.fps, qcam_info.bitrate,
-                                       qcam_info.is_h265 ? cereal::EncodeIndex::Type::FULL_H_E_V_C : cereal::EncodeIndex::Type::QCAMERA_H264,
-                                       qcam_info.frame_width, qcam_info.frame_height, true));
-      }
-    }
-
-    while (!do_exit) {
-      VisionIpcBufExtra extra;
-      VisionBuf* buf = vipc_client.recv(&extra);
-      if (buf == nullptr) continue;
-
-      if (cam_info.trigger_rotate) {
-        s->last_camera_seen_tms = millis_since_boot();
-        if (!sync_encoders(s, cam_info.type, extra.frame_id)) {
-          continue;
-        }
-
-        // check if we're ready to rotate
-        trigger_rotate_if_needed(s, cur_seg, extra.frame_id);
-        if (do_exit) break;
-      }
-
-      // rotate the encoder if the logger is on a newer segment
-      if (s->rotate_segment > cur_seg) {
-        cur_seg = s->rotate_segment;
-
-        LOGW("camera %d rotate encoder to %s", cam_info.type, s->segment_path);
-        for (auto &e : encoders) {
-          e->encoder_close();
-          e->encoder_open(s->segment_path);
-        }
-        if (lh) {
-          lh_close(lh);
-        }
-        lh = logger_get_handle(&s->logger);
-      }
-
-      // encode a frame
-      for (int i = 0; i < encoders.size(); ++i) {
-        int out_id = encoders[i]->encode_frame(buf->y, buf->u, buf->v,
-                                               buf->width, buf->height, &extra);
-
-        if (out_id == -1) {
-          LOGE("Failed to encode frame. frame_id: %d encode_id: %d", extra.frame_id, encode_idx);
-        }
-
-        // publish encode index
-        if (i == 0 && out_id != -1) {
-          MessageBuilder msg;
-          // this is really ugly
-          bool valid = (buf->get_frame_id() == extra.frame_id);
-          auto eidx = cam_info.type == DriverCam ? msg.initEvent(valid).initDriverEncodeIdx() :
-                     (cam_info.type == WideRoadCam ? msg.initEvent(valid).initWideRoadEncodeIdx() : msg.initEvent(valid).initRoadEncodeIdx());
-          eidx.setFrameId(extra.frame_id);
-          eidx.setTimestampSof(extra.timestamp_sof);
-          eidx.setTimestampEof(extra.timestamp_eof);
-          if (Hardware::TICI()) {
-            eidx.setType(cereal::EncodeIndex::Type::FULL_H_E_V_C);
-          } else {
-            eidx.setType(cam_info.type == DriverCam ? cereal::EncodeIndex::Type::FRONT : cereal::EncodeIndex::Type::FULL_H_E_V_C);
-          }
-          eidx.setEncodeId(encode_idx);
-          eidx.setSegmentNum(cur_seg);
-          eidx.setSegmentId(out_id);
-          if (lh) {
-            auto bytes = msg.toBytes();
-            lh_log(lh, bytes.begin(), bytes.size(), true);
-          }
-          pm.send(service_name, msg);
-        }
-      }
-
-      encode_idx++;
-    }
-
-    if (lh) {
-      lh_close(lh);
-      lh = NULL;
-    }
-  }
-
-  LOG("encoder destroy");
-  for(auto &e : encoders) {
-    e->encoder_close();
-    delete e;
-  }
-}
+struct LoggerdState {
+  LoggerState logger = {};
+  char segment_path[4096];
+  std::mutex rotate_lock;
+  std::atomic<int> rotate_segment;
+  std::atomic<double> last_camera_seen_tms;
+  std::atomic<int> ready_to_rotate;  // count of encoders ready to rotate
+  int max_waiting = 0;
+  double last_rotate_tms = 0.;      // last rotate time in ms
+};
 
 void logger_rotate(LoggerdState *s) {
   {
@@ -169,7 +24,6 @@ void logger_rotate(LoggerdState *s) {
     s->ready_to_rotate = 0;
     s->last_rotate_tms = millis_since_boot();
   }
-  s->rotate_cv.notify_all();
   LOGW((s->logger.part == 0) ? "logging to %s" : "rotated to %s", s->segment_path);
 }
 
@@ -187,6 +41,99 @@ void rotate_if_needed(LoggerdState *s) {
   }
 }
 
+struct RemoteEncoder {
+  std::unique_ptr<VideoWriter> writer;
+  int segment = -1;
+  std::vector<Message *> q;
+  int logger_segment = -1;
+  int dropped_frames = 0;
+  bool recording = false;
+};
+
+int handle_encoder_msg(LoggerdState *s, Message *msg, std::string &name, struct RemoteEncoder &re) {
+  const LogCameraInfo &cam_info = (name == "driverEncodeData") ? cameras_logged[1] :
+    ((name == "wideRoadEncodeData") ? cameras_logged[2] :
+    ((name == "qRoadEncodeData") ? qcam_info : cameras_logged[0]));
+
+  // rotation happened, process the queue (happens before the current message)
+  int bytes_count = 0;
+  if (re.logger_segment != s->rotate_segment) {
+    re.logger_segment = s->rotate_segment;
+    for (auto &qmsg: re.q) {
+      bytes_count += handle_encoder_msg(s, qmsg, name, re);
+    }
+    re.q.clear();
+  }
+
+  // extract the message
+  capnp::FlatArrayMessageReader cmsg(kj::ArrayPtr<capnp::word>((capnp::word *)msg->getData(), msg->getSize()));
+  auto event = cmsg.getRoot<cereal::Event>();
+  auto edata = (name == "driverEncodeData") ? event.getDriverEncodeData() :
+    ((name == "wideRoadEncodeData") ? event.getWideRoadEncodeData() :
+    ((name == "qRoadEncodeData") ? event.getQRoadEncodeData() : event.getRoadEncodeData()));
+  auto idx = edata.getIdx();
+  auto flags = idx.getFlags();
+
+  if (!re.recording) {
+    // only create on iframe
+    if (flags & V4L2_BUF_FLAG_KEYFRAME) {
+      if (re.dropped_frames) {
+        // this should only happen for the first segment, maybe
+        LOGD("%s: dropped %d non iframe packets before init", name.c_str(), re.dropped_frames);
+        re.dropped_frames = 0;
+      }
+      // if we aren't recording, don't create the writer
+      if (cam_info.record) {
+        re.writer.reset(new VideoWriter(s->segment_path,
+          cam_info.filename, idx.getType() != cereal::EncodeIndex::Type::FULL_H_E_V_C,
+          cam_info.frame_width, cam_info.frame_height, cam_info.fps, idx.getType()));
+        // write the header
+        auto header = edata.getHeader();
+        re.writer->write((uint8_t *)header.begin(), header.size(), idx.getTimestampEof()/1000, true, false);
+      }
+      re.segment = idx.getSegmentNum();
+      re.recording = true;
+    } else {
+      ++re.dropped_frames;
+      return bytes_count;
+    }
+  }
+
+  if (re.segment != idx.getSegmentNum()) {
+    if (re.recording) {
+      // encoder is on the next segment, this segment is over so we close the videowriter
+      re.writer.reset();
+      re.recording = false;
+      ++s->ready_to_rotate;
+      LOGD("rotate %d -> %d ready %d/%d for %s", re.segment, idx.getSegmentNum(), s->ready_to_rotate.load(), s->max_waiting, name.c_str());
+    }
+    // queue up all the new segment messages, they go in after the rotate
+    re.q.push_back(msg);
+  } else {
+    if (re.writer) {
+      auto data = edata.getData();
+      re.writer->write((uint8_t *)data.begin(), data.size(), idx.getTimestampEof()/1000, false, flags & V4L2_BUF_FLAG_KEYFRAME);
+    }
+
+    // put it in log stream as the idx packet
+    MessageBuilder bmsg;
+    auto evt = bmsg.initEvent(event.getValid());
+    evt.setLogMonoTime(event.getLogMonoTime());
+    if (name == "driverEncodeData") { evt.setDriverEncodeIdx(idx); }
+    if (name == "wideRoadEncodeData") { evt.setWideRoadEncodeIdx(idx); }
+    if (name == "qRoadEncodeData") { evt.setQRoadEncodeIdx(idx); }
+    if (name == "roadEncodeData") { evt.setRoadEncodeIdx(idx); }
+    auto new_msg = bmsg.toBytes();
+    logger_log(&s->logger, (uint8_t *)new_msg.begin(), new_msg.size(), true);   // always in qlog?
+    bytes_count += new_msg.size();
+
+    // this frees the message
+    delete msg;
+  }
+
+  return bytes_count;
+}
+
 void loggerd_thread() {
   // setup messaging
   typedef struct QlogState {
@@ -202,7 +149,7 @@ void loggerd_thread() {
 
   // subscribe to all socks
   for (const auto& it : services) {
-    const bool encoder = env_remote_encoder && (strcmp(it.name+strlen(it.name)-strlen("EncodeData"), "EncodeData") == 0);
+    const bool encoder = strcmp(it.name+strlen(it.name)-strlen("EncodeData"), "EncodeData") == 0;
     if (!it.should_log && !encoder) continue;
     LOGD("logging %s (on port %d)", it.name, it.port);
 
@@ -225,15 +172,10 @@ void loggerd_thread() {
 
   // init encoders
   s.last_camera_seen_tms = millis_since_boot();
-  std::vector<std::thread> encoder_threads;
   for (const auto &cam : cameras_logged) {
     if (cam.enable) {
-      if (env_remote_encoder) {
-        if (cam.has_qcamera) { s.max_waiting++; }
-      } else {
-        encoder_threads.push_back(std::thread(encoder_thread, &s, cam));
-      }
-      if (cam.trigger_rotate) s.max_waiting++;
+      s.max_waiting++;
+      if (cam.has_qcamera) { s.max_waiting++; }
     }
   }
 
@@ -252,6 +194,7 @@ void loggerd_thread() {
         const bool in_qlog = qs.freq != -1 && (qs.counter++ % qs.freq == 0);
 
         if (qs.encoder) {
+          s.last_camera_seen_tms = millis_since_boot();
           bytes_count += handle_encoder_msg(&s, msg, qs.name, remote_encoders[sock]);
         } else {
           logger_log(&s.logger, (uint8_t *)msg->getData(), msg->getSize(), in_qlog);
@@ -275,10 +218,6 @@ void loggerd_thread() {
     }
   }
 
-  LOGW("closing encoders");
-  s.rotate_cv.notify_all();
-  for (auto &t : encoder_threads) t.join();
-
   LOGW("closing logger");
   logger_close(&s.logger, &do_exit);
 
@@ -290,4 +229,19 @@ void loggerd_thread() {
 
   // messaging cleanup
   for (auto &[sock, qs] : qlog_states) delete sock;
+}
+
+int main(int argc, char** argv) {
+  if (Hardware::TICI()) {
+    int ret;
+    ret = util::set_core_affinity({0, 1, 2, 3});
+    assert(ret == 0);
+    // TODO: why does this impact camerad timings?
+    //ret = util::set_realtime_priority(1);
+    //assert(ret == 0);
+  }
+
+  loggerd_thread();
+
+  return 0;
 }
